Moves errorType to converter
